% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_sinew.R
\name{init_sinew}
\alias{init_sinew}
\title{init_sinew}
\usage{
init_sinew(add_fields = c("details", "examples", "rdname", "export",
  "author"), author = as.character(person("Lorenzo", "Busetto, phD (2017)",
  email = "lbusett@gmail.com")))
}
\arguments{
\item{add_fields}{PARAM_DESCRIPTION, Default: c("details", "examples", "rdname", "export", "author")}

\item{author}{PARAM_DESCRIPTION, Default: as.character(person("Lorenzo", "Busetto, phD (2017)", email = "lbusett@gmail.com"))}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Lorenzo Busetto, phD (2017) <lbusett@gmail.com>
}
