% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_linregr.R
\name{lb_linregr}
\alias{lb_linregr}
\title{lb_linregr}
\usage{
lb_linregr(in_data, X_VAR, Y_VAR, GP_VAR = NULL, na.rm = TRUE)
}
\arguments{
\item{in_data}{PARAM_DESCRIPTION}

\item{X_VAR}{PARAM_DESCRIPTION}

\item{Y_VAR}{PARAM_DESCRIPTION}

\item{GP_VAR}{PARAM_DESCRIPTION, Default: NULL}

\item{na.rm}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Lorenzo Busetto, phD (2017) <lbusett@gmail.com>
}
