% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_plotregr.R
\name{lb_plotregr}
\alias{lb_plotregr}
\title{FUNCTION_TITLE}
\usage{
lb_plotregr(in_data, GP_VAR, X_VAR, Y_VAR, common_range = TRUE,
  title = NULL, subtitle = NULL)
}
\arguments{
\item{in_data}{PARAM_DESCRIPTION}

\item{GP_VAR}{PARAM_DESCRIPTION}

\item{X_VAR}{PARAM_DESCRIPTION}

\item{Y_VAR}{PARAM_DESCRIPTION}

\item{common_range}{PARAM_DESCRIPTION, Default: TRUE}

\item{title}{PARAM_DESCRIPTION, Default: NULL}

\item{subtitle}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Lorenzo Busetto, phD (2017) <lbusett@gmail.com>
}
