% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_sinew.R
\name{init_sinew}
\alias{init_sinew}
\title{Initialize SINEW opts for roxygenize}
\usage{
init_sinew()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Lorenzo Busetto, phD (2017) <lbusett@gmail.com>
}
